docs: add site demo
