+++ packages/site/examples/interaction/select/demo/meta.json
@@ -5,7 +5,7 @@
   },
   "demos": [
     {
-      "filename": "click.js",
+      "filename": "click.ts",
       "title": "点击单选或多选",
       "screenshot": "https://gw.alipayobjects.com/mdn/rms_f8c6a0/afts/img/A*F0RGQ67xbGYAAAAAAAAAAAAAARQnAQ"
     },
